Refactor outdated SignalSlot implementation to Events
