+++ Classes/EventListener/AfterFileUpdatedInIndex.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace B3N\AzureStorage\TYPO3\EventListener;
+
+use TYPO3\CMS\Core\Resource\Event\AfterFileUpdatedInIndexEvent;
+
+class AfterFileUpdatedInIndex extends AbstractFileIndexEventListener
+{
+    public function __invoke(AfterFileUpdatedInIndexEvent $event): void
+    {
+        $this->recordUpdatedOrCreated($event->getRelevantProperties());
+    }
+}
